Move download of technology-data
