+++ src/egon/data/datasets/scenario_parameters/__init__.py
@@ -8,6 +8,12 @@ from sqlalchemy.dialects.postgresql import JSONB
 import pandas as pd
 import egon.data.datasets.scenario_parameters.parameters as parameters
 from egon.data.datasets import Dataset
+from urllib.request import urlretrieve
+import zipfile
+
+import egon.data.config
+from pathlib import Path
+import shutil
 
 Base = declarative_base()
 
@@ -157,12 +163,32 @@ def get_sector_parameters(sector, scenario=None):
 
     return values
 
+def download_pypsa_technology_data():
+    """Downlad PyPSA technology data results."""
+    data_path = Path(".") / "pypsa_technology_data"
+    # Delete folder if it already exists
+    if data_path.exists() and data_path.is_dir():
+        shutil.rmtree(data_path)
+    # Get parameters from config and set download URL
+    sources = egon.data.config.datasets()["pypsa-technology-data"]["sources"][
+        "zenodo"
+    ]
+    url = f"""https://zenodo.org/record/{sources['deposit_id']}/files/PyPSA/technology-data-v0.3.0.zip"""
+    target_file = egon.data.config.datasets()["pypsa-technology-data"][
+        "targets"
+    ]["file"]
 
+    # Retrieve files
+    urlretrieve(url, target_file)
+
+    with zipfile.ZipFile(target_file, "r") as zip_ref:
+        zip_ref.extractall(".")
+        
 class ScenarioParameters(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="ScenarioParameters",
-            version="0.0.2",
+            version="0.0.3",
             dependencies=dependencies,
-            tasks=(create_table, insert_scenarios),
+            tasks=(create_table, download_pypsa_technology_data, insert_scenarios),
         )
