+++ src/egon/data/datasets/scenario_parameters/parameters.py
@@ -169,41 +169,3 @@ def heat(scenario):
         print(f"Scenario name {scenario} is not valid.")
 
     return parameters
-
-
-def download_pypsa_technology_data():
-    """Downlad PyPSA technology data results."""
-    data_path = Path(".") / "pypsa_technology_data"
-    # Delete folder if it already exists
-    if data_path.exists() and data_path.is_dir():
-        shutil.rmtree(data_path)
-    # Get parameters from config and set download URL
-    sources = egon.data.config.datasets()["pypsa-technology-data"]["sources"]["zenodo"]
-    url = f"""https://zenodo.org/record/{sources['deposit_id']}/files/technology-data-v0.3.0.zip"""
-    target_file = egon.data.config.datasets()["pypsa-technology-data"]["targets"]["file"]
-
-    # Retrieve files
-    urlretrieve(url, target_file)
-
-    with zipfile.ZipFile(target_file, "r") as zip_ref:
-        zip_ref.extractall(".")
-
-
-def insert_pypsa_technology_data(scenario):
-    """Insert the technology data from pypsa into the scenario table.
-
-    Parameters
-    ----------
-    scenario : str
-        Name of the scenario.
-    """
-    if scenario == "eGon2035":
-        data = '2035'
-
-    elif scenario == "eGon100RE":
-        data = '2050'
-
-    else:
-        print(f"Scenario name {scenario} is not valid.")
-    #do_stuff
-    return
